perf: size the general mixed kernel to its operands, and merge coverage backends
Two things the last review left open.

**The 1024-bit general kernel.** Every cross-scale operation that missed the
narrow mixed_native path did all of its arithmetic in sixteen 64-bit limbs, no
matter how small the operands were. A Fixed64 x Fixed64 multiply into a
Fixed128 needs about eighty bits and was running a 16x16 schoolbook multiply:
256 partial products, of which four were not multiplying by zero. pow10_limbs
built its powers of ten in 1024 bits by repeated shift-add, and divmod_knuth
normalised and looped over all sixteen limbs of operands that occupied two.

src/mixed.cpp now computes an upper bound on the bits the numerator and
denominator actually need -- operand widths plus ceil(e * log2 10) for each
power of ten -- and dispatches to the smallest of four tiers: 128, 256, 512 or
1024 bits. The bound must never under-estimate, so pow10_bits uses e*10/3 + 1,
which sits above log2(10) for every e.

  mul_to, 64-bit operands   8561 -> 416   instructions/op
  div_to, 64-bit operands   8377 -> 729
  mul_to, 256-bit operands  8561 -> 1764

The 1024-bit path still exists and is still correct; it is now reached only by
operands that need it. Same results throughout: 1,017,500 differential checks
against Boost.Multiprecision unchanged across the native and portable backends,
34/34 tests in native, portable and no-__int128 configurations, and clean under
ASan and UBSan on both backends. Every other workload in the instruction-count
baseline is byte-identical, which is the evidence that the change is confined to
this kernel.

Three benchmark rows were added for 256-bit operands, because without them the
8- and 16-limb tiers would not have been gated at all.

The cliff is now a gradient: 53 / 416 / 1764 for a multiply, against a flat 8561
for anything off the native path before. What is left is real work -- the
general path evaluates an exact rational and performs a division the native path
avoids -- and the docs say so rather than calling it a cliff.

**Coverage measured one backend.** src/division.hpp read 10.32% line coverage,
which looked like a testing gap and was a measurement bug: the portable and
no-int128 jobs exercise it thoroughly, and a native-only profile cannot see
that. The job now builds both backends, runs both suites, and merges the
profiles. division.hpp reads 98.67%; the totals move from 72.1%/77.5% to
75.1%/77.9% locally. The threshold is a ratchet set below the first merged CI
measurement rather than guessed at, and STATUS.md says to raise it and never
lower it.

Also fixes a latent signed/unsigned conversion in uint_limbs::bit_width that
only surfaced once the template was instantiated at the new widths.

+++ src/limbs.hpp
@@ -54,7 +54,8 @@ struct uint_limbs {
     [[nodiscard]] constexpr unsigned bit_width() const noexcept {
         for (std::size_t i = L; i > 0; --i) {
             if (limbs[i - 1] != 0) {
-                return static_cast<unsigned>(i * 64 - std::countl_zero(limbs[i - 1]));
+                // countl_zero returns int; do the subtraction in one signedness.
+                return static_cast<unsigned>(i * 64) - static_cast<unsigned>(std::countl_zero(limbs[i - 1]));
             }
         }
         return 0;
